backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-multipart
httpx
pydantic-settings
bcrypt
python-jose[cryptography]
google-generativeai
